products/serializers.py
from products.models import Product
from rest_framework import serializers


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id', 'name', 'quantity','unique_id']
        read_only_fields = ['unique_id']
